# Optimize uploading ballots for recreating prior elections