--- a/packages/backend/src/Models/IBallotStore.ts
+++ b/packages/backend/src/Models/IBallotStore.ts
@@ -4,6 +4,7 @@ import { ILoggingContext } from "../Services/Logging/ILogger";
 
 export interface IBallotStore {
     submitBallot: (ballot: Ballot, ctx:ILoggingContext, reason:string) => Promise<Ballot>;
+    bulkSubmitBallots: (ballots: Ballot[], ctx:ILoggingContext, reason:string) => Promise<Ballot[]>;
     getBallotByID: (ballot_id: string, ctx:ILoggingContext) => Promise<Ballot | null>;
     getBallotsByElectionID: (election_id: string,  ctx:ILoggingContext) => Promise<Ballot[] | null>;
     delete(ballot_id: Uid,  ctx:ILoggingContext, reason:string): Promise<boolean>;