--- a/packages/shared/src/domain_model/ITabulators.ts
+++ b/packages/shared/src/domain_model/ITabulators.ts
@@ -1,4 +1,4 @@
-export type score = number
+export type score = number | null
 
 export type ballot = score[]