--- a/packages/shared/src/domain_model/ITabulators.ts
+++ b/packages/shared/src/domain_model/ITabulators.ts
@@ -2,6 +2,8 @@ export type score = number | null
 
 export type ballot = score[]
 
+export type nonNullBallot = number[]
+
 export type ballots = ballot[]
 
 export interface candidate {