--- a/packages/shared/src/domain_model/Vote.ts
+++ b/packages/shared/src/domain_model/Vote.ts
@@ -5,3 +5,6 @@ export interface Vote {
     race_id: Uid;        // Must match the pollId of the election
     scores: Score[];       // One per candidate
 }
+
+// this format is used in bulk uploads where the race/candidate order is mapped in a separate structure
+export type OrderedVote = number[];