… packages/backend/src/OpenApi/swagger.json
… packages/backend/src/Tabulators/AllocatedScore.test.ts
… packages/backend/src/Tabulators/Approval.test.ts
… packages/backend/src/Tabulators/Plurality.test.ts
… packages/backend/src/Tabulators/Star.test.ts